requirements.txt
mcp >=0.1.0
python-dotenv >=1.0.0
loguru >=0.7.0
pydantic >=2.0.0
typing-extensions >=4.0.0
anthropic>=0.8.0
openai>=1.0.0
